Move tests of internal API to corresponding file
